Rather hairy code to fix indentation.
Basically, PICASM-LINE-PIECES analyses a line and breaks it into
pieces for re-indentation. Then PICASM-INDENT-LINE actually does the
work. We also no longer indent by tabbing, but rather with goal
columns, using INDENT-TO to work correctly with
INDENT-TABS-MODE (hopefully).

The result seems to correctly indent a largish hunk of assembly, so I
*think* it's correct!